fixes
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@419 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/client/CClient_Parse.cpp
@@ -471,7 +471,7 @@ bool CClient::ParsePrepareGame(CBytestream *bs)
 	if (lv)  {
 		lv->Clear();
 		line_t *l = NULL;
-		for (uint i=MAX((uint)0,cChatbox.getNumLines()-254);i<cChatbox.getNumLines();i++)  {
+		for (uint i=(uint)MAX((int)0,(int)(cChatbox.getNumLines()-254));i<cChatbox.getNumLines();i++)  {
 			l = cChatbox.GetLine(i);
 			// Add only chat text
 			if (l) if (l->iColour == tLX->clChatText)  {
