+++ src/client/Menu_Net_Host.cpp
@@ -606,7 +606,7 @@ void Menu_Net_HostGotoLobby(void)
 	CListview *lv = (CListview *)cHostLobby.getWidget(hl_ChatList);
 	if (lv)  {
 		line_t *l = NULL;
-		for (uint i=MAX((uint)0,cClient->getChatbox()->getNumLines()-255);i<cClient->getChatbox()->getNumLines();i++)  {
+		for (uint i=(uint)MAX((int)0,(int)(cClient->getChatbox()->getNumLines()-255));i<cClient->getChatbox()->getNumLines();i++)  {
 			l = cClient->getChatbox()->GetLine(i);
 			if (l) if (l->iColour == tLX->clChatText)  {
 				if(lv->getLastItem())
