+++ src/client/Menu_Net_Join.cpp
@@ -538,7 +538,7 @@ void Menu_Net_JoinGotoLobby(void)
 	CListview *lv = (CListview *)cJoinLobby.getWidget(jl_ChatList);
 	if (lv)  {
 		line_t *l = NULL;
-		for (uint i=MAX((uint)0,cClient->getChatbox()->getNumLines()-255);i<cClient->getChatbox()->getNumLines();i++)  {
+		for (uint i=(uint)MAX((int)0,(int)(cClient->getChatbox()->getNumLines()-255));i<cClient->getChatbox()->getNumLines();i++)  {
 			l = cClient->getChatbox()->GetLine(i);
 			if (l) if (l->iColour == tLX->clChatText)  {
 				if(lv->getLastItem())
